Update several dependencies in composer.lock file
The project dependencies have been updated in the composer.lock file. This includes minor and patch version upgrades of several packages such as drupal/coder, phpstan/phpdoc-parser, squizlabs/php_codesniffer, and symfony related packages. Additionally, an adjustment was made to the image style configuration and a corresponding update hook was added in the osu_media module.